% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodestar_console.R
\name{lodestar}
\alias{lodestar}
\title{Lodestar controller - for calling from Rscript}
\usage{
lodestar()
}
\value{
nothing much
}
\description{
A load of the Lodestar code is intended to be run through NextFlow scripts;
this function provides some control infrastructure for the usage of the
package through the commandline in so far is possible using R
e.g. \verb{Rscript -e 'gwangmyeongseong3::lodestar()'}
}
