% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodestar.R
\name{lodestar_creds}
\alias{lodestar_creds}
\title{Inject a username and password into the backend keyring environment}
\usage{
lodestar_creds(
  username,
  password,
  service = .default_service,
  keyring = .default_keyring
)
}
\arguments{
\item{username}{username}

\item{password}{password}

\item{service}{the database to use}

\item{keyring}{the keyring where credentials will be stored}
}
\value{
backend instance
}
\description{
We are working here to prepare a clean system for the simple and ad hoc
analysis of high dimensional sequence data. This accessory method is intended
to help define the limits of RDBMS, database, username and password required
for connecting to the most appropriate systems.
}
