% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodestar_console.R
\name{lodestar_upload}
\alias{lodestar_upload}
\title{#' @importFrom yaml write_yaml
lodestar_make_yaml = function(opt) {
cli::cli_h1("creating YAML config file")
cli::cli_alert_info(paste0("using \link{",opt$config,"} as YAML destination"))
lsc <- LodestarConn$new(
keyring=opt$rdbms,
service=opt$database,
username=opt$username,
password=opt$password)
yaml::write_yaml(lsc$as_yaml(), opt$config)
cli::cli_alert_success(paste0("\link{",opt$config,"} written successfully"))
}}
\usage{
lodestar_upload(opt)
}
\description{
#' @importFrom yaml write_yaml
lodestar_make_yaml = function(opt) {
cli::cli_h1("creating YAML config file")
cli::cli_alert_info(paste0("using \link{",opt$config,"} as YAML destination"))
lsc <- LodestarConn$new(
keyring=opt$rdbms,
service=opt$database,
username=opt$username,
password=opt$password)
yaml::write_yaml(lsc$as_yaml(), opt$config)
cli::cli_alert_success(paste0("\link{",opt$config,"} written successfully"))
}
}
