% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodestar.R
\name{lodestar_user_tibble}
\alias{lodestar_user_tibble}
\title{Prepare a summary of available keyring annotations}
\usage{
lodestar_user_tibble()
}
\value{
a tibble
}
\description{
In a larger SQL environment there is a possibility of 10s of different
RDBMS systems and considerable numbre of database instances - this simple
method prepares a tibble that summarises the available keyring, service and
username information for reminding the available connection possibilities
}
