% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodestar.R
\name{LodestarConn}
\alias{LodestarConn}
\title{R6 Class for abstracting database connections and maintaining lodestar schema}
\description{
This class aims to provide some basic minimal functionality for interacting
with a Lodestar schema in a relational database. A key ambition at outset
has been the separation of user credentials and demonstration code for the
preparation of examples and documentation in the vignettes.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LodestarConn$new()}}
\item \href{#method-connection}{\code{LodestarConn$connection()}}
\item \href{#method-as_yaml}{\code{LodestarConn$as_yaml()}}
\item \href{#method-clone}{\code{LodestarConn$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new LodestarConn object. This
initialisation method orchestrates other sanity checking
of the defined parameters(s) to ensure that the environment is coherent
and tractable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LodestarConn$new(
  backend = NA,
  keyring = .default_keyring,
  service = .default_service,
  username = NA,
  password = NA,
  port = 5432,
  silent = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{backend}}{keyring backend object used to define username and
password independently of the vignette code (persistence) - NA by default
\itemize{
\item software will endeavour to select backend on basis of available
keyring context
}}

\item{\code{keyring}}{which of the available keyrings will hold the connection
details - \code{lodestar} by default.}

\item{\code{service}}{the name of the database that we'll endeavour to connect
to - this will be parsed from the backend context - \code{lodestar} by
default.}

\item{\code{username}}{the user account to be used with the database - NA by
default; in simpler installations the software will identify the username
on basis of e.g. service}

\item{\code{password}}{the password to connect to the database with; using the
backend object this can be read from the keyring}

\item{\code{silent}}{boolean defining whether logging of process should be
performed (TRUE by default)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the LodestarConn R6 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connection"></a>}}
\if{latex}{\out{\hypertarget{method-connection}{}}}
\subsection{Method \code{connection()}}{
Get a database connection corresponding to the passed parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LodestarConn$connection()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a DBI connection object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_yaml"></a>}}
\if{latex}{\out{\hypertarget{method-as_yaml}{}}}
\subsection{Method \code{as_yaml()}}{
export the key characteristics of this object in a format for writing
to YAML to allow the propagation of settings in a cluster environment
where we cannot expect processes to have access to the keyring - this
does mean that connection strings and passwords are written to file ...
this will be addressed more securely once we're beyond the proof-of-
concept.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LodestarConn$as_yaml()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LodestarConn$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
